+++ pycbc/inference/__init__.py
@@ -2,4 +2,4 @@ from pycbc.inference.distributions import *
 from pycbc.inference.likelihood import *
 from pycbc.inference.sampler import *
 from pycbc.inference.prior import *
-from pycbc.inference.boundary_utils import *
+from pycbc.inference.boundaries import *
