+++ pycbc/inference/boundaries.py
